# swagger.yaml
openapi: 3.0.0
info:
  title: DiscoverIT API
  description: Node.js + MongoDB - DiscoverIT API with accounts and nagios monitor
  version: 1.0.0
servers:
  - url: http://localhost:70
    description: Local development server
  - url: http://169.59.166.94:70
    description: cloud development server
paths:
...
  /tickets:
    get:
      summary: get array of all tickets
      description: users can see tickets
      operationId: getTickets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: list of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: "string"
                    example: "xxxxxxxxxxxxxxxxx"
                  ticketID:
                    type: string
                    example: "client-support-srv-1"
                  equipment:
                    type: string
                    example: "server"
                  model:
                    type: string
                    example: "5000x"
                  title:
                    type: string
                    example: "soporte equipo"
                  description:
                    type: string
                    example: "falla en disco"
                  severity:
                    type: string
                    example: "critico"
                  client:
                    type: string
                    example: "virwo"
                  issueType:  
                    type: string
                    example: "hardware"
                  userName:  
                    type: string
                    example: "Juan Escotia"
                  contract:  
                    type: string
                    example: "p-226845"
                  email:  
                    type: string
                    example: "juan.escotia@virwo.com"
                  telephone:  
                    type: string
                    example: "55555555555"
                  equipmentSerial:  
                    type: string
                    example: "58dfa5d"
                  clientEvidencePath:  
                    type: string
                    example: "/client/contract/ticket"
                  assignedSupportOperator:  
                    type: string
                    example: "barabas"  
                  status:  
                    type: string
                    example: "pending"
                  created:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  responseComments:  
                    type: string
                    example: "barabas"
                  responseDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  closeDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  supportEvidencePath:  
                    type: string
                    example: "/client/contract/ticket-support"  
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new contract and equipment
      description: Restricted to admin users.
      operationId: createTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  equipment:
                    type: string
                    example: "server"
                  model:
                    type: string
                    example: "5000x"
                  title:
                    type: string
                    example: "soporte equipo"
                  description:
                    type: string
                    example: "falla en disco"
                  severity:
                    type: string
                    example: "critico"
                  client:
                    type: string
                    example: "virwo"
                  issueType:  
                    type: string
                    example: "hardware"
                  userName:  
                    type: string
                    example: "Juan Escotia"
                  contract:  
                    type: string
                    example: "p-226845"
                  email:  
                    type: string
                    example: "juan.escotia@virwo.com"
                  telephone:  
                    type: string
                    example: "55555555555"
                  equipmentSerial:  
                    type: string
                    example: "58dfa5d"
                  clientEvidencePath:  
                    type: string
                    example: "/client/contract/ticket"
            
              required:
                - equipment
                - model
                - title
                - description
                - severity
                - client
                - issueType
                - userName
                - contract
                - End
                - email
                - telephone
                - equipmentSerial
      responses:
        "200":
          description: ticket created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: "string"
                    example: "xxxxxxxxxxxxxxxxx"
                  ticketID:
                    type: string
                    example: "client-support-srv-1"
                  equipment:
                    type: string
                    example: "server"
                  model:
                    type: string
                    example: "5000x"
                  title:
                    type: string
                    example: "soporte equipo"
                  description:
                    type: string
                    example: "falla en disco"
                  severity:
                    type: string
                    example: "critico"
                  client:
                    type: string
                    example: "virwo"
                  issueType:  
                    type: string
                    example: "hardware"
                  userName:  
                    type: string
                    example: "Juan Escotia"
                  contract:  
                    type: string
                    example: "p-226845"
                  email:  
                    type: string
                    example: "juan.escotia@virwo.com"
                  telephone:  
                    type: string
                    example: "55555555555"
                  equipmentSerial:  
                    type: string
                    example: "58dfa5d"
                  clientEvidencePath:  
                    type: string
                    example: "/client/contract/ticket"
                  assignedSupportOperator:  
                    type: string
                    example: "barabas"  
                  status:  
                    type: string
                    example: "pending"
                  created:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  responseComments:  
                    type: string
                    example: "barabas"
                  responseDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  closeDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  supportEvidencePath:  
                    type: string
                    example: "/client/contract/ticket-support"  
        "400":
          description: Serial number is already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Serial number is already registered"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /tickets/{id}:
    parameters:
      - in: path
        name: id
        description: ticket object id
        required: true
        example: "5eb12e197e06a76ccdefc121"
        schema:
          type: string
    get:
      summary: get ticket information with id
      description: get ticket infor
      operationId: getTicketsById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified ticket
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: "string"
                    example: "xxxxxxxxxxxxxxxxx"
                  ticketID:
                    type: string
                    example: "client-support-srv-1"
                  equipment:
                    type: string
                    example: "server"
                  model:
                    type: string
                    example: "5000x"
                  title:
                    type: string
                    example: "soporte equipo"
                  description:
                    type: string
                    example: "falla en disco"
                  severity:
                    type: string
                    example: "critico"
                  client:
                    type: string
                    example: "virwo"
                  issueType:  
                    type: string
                    example: "hardware"
                  userName:  
                    type: string
                    example: "Juan Escotia"
                  contract:  
                    type: string
                    example: "p-226845"
                  email:  
                    type: string
                    example: "juan.escotia@virwo.com"
                  telephone:  
                    type: string
                    example: "55555555555"
                  equipmentSerial:  
                    type: string
                    example: "58dfa5d"
                  clientEvidencePath:  
                    type: string
                    example: "/client/contract/ticket"
                  assignedSupportOperator:  
                    type: string
                    example: "barabas"  
                  status:  
                    type: string
                    example: "pending"
                  created:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  responseComments:  
                    type: string
                    example: "barabas"
                  responseDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  closeDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  supportEvidencePath:  
                    type: string
                    example: "/client/contract/ticket-support"  
                  
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
          
    put:
      summary: Update a ticket
      description: admin user can update a ticket
      operationId: updateTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  equipment:
                    type: string
                    example: "server"
                  model:
                    type: string
                    example: "5000x"
                  title:
                    type: string
                    example: "soporte equipo"
                  description:
                    type: string
                    example: "falla en disco"
                  severity:
                    type: string
                    example: "critico"
                  client:
                    type: string
                    example: "virwo"
                  issueType:  
                    type: string
                    example: "hardware"
                  userName:  
                    type: string
                    example: "Juan Escotia"
                  contract:  
                    type: string
                    example: "p-226845"
                  email:  
                    type: string
                    example: "juan.escotia@virwo.com"
                  telephone:  
                    type: string
                    example: "55555555555"
                  equipmentSerial:  
                    type: string
                    example: "58dfa5d"
                  clientEvidencePath:  
                    type: string
                    example: "/client/contract/ticket"
                  assignedSupportOperator:  
                    type: string
                    example: "barabas"  
                  status:  
                    type: string
                    example: "pending"
                  responseComments:  
                    type: string
                    example: "barabas"
                  responseDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  closeDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  supportEvidencePath:  
                    type: string
                    example: "/client/contract/ticket-support"  
      responses:
        "200":
          description: ticket updated successfully. The details of the updated ticket are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: "string"
                    example: "xxxxxxxxxxxxxxxxx"
                  ticketID:
                    type: string
                    example: "client-support-srv-1"
                  equipment:
                    type: string
                    example: "server"
                  model:
                    type: string
                    example: "5000x"
                  title:
                    type: string
                    example: "soporte equipo"
                  description:
                    type: string
                    example: "falla en disco"
                  severity:
                    type: string
                    example: "critico"
                  client:
                    type: string
                    example: "virwo"
                  issueType:  
                    type: string
                    example: "hardware"
                  userName:  
                    type: string
                    example: "Juan Escotia"
                  contract:  
                    type: string
                    example: "p-226845"
                  email:  
                    type: string
                    example: "juan.escotia@virwo.com"
                  telephone:  
                    type: string
                    example: "55555555555"
                  equipmentSerial:  
                    type: string
                    example: "58dfa5d"
                  clientEvidencePath:  
                    type: string
                    example: "/client/contract/ticket"
                  assignedSupportOperator:  
                    type: string
                    example: "barabas"  
                  status:  
                    type: string
                    example: "pending"
                  created:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  responseComments:  
                    type: string
                    example: "barabas"
                  responseDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  closeDate:
                    type: string
                    example: "2020-05-08T03:11:21.553Z"
                  supportEvidencePath:  
                    type: string
                    example: "/client/contract/ticket-support"  
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"